Renamed Folder to styles
