+++ front-end/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import logo from './logo.svg';
-import './scss/App.scss';
+import './styles/App.scss';
 
 class App extends Component {
   render() {
