+++ front-end/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './scss/index.scss';
+import './styles/index.scss';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
