+++ front-end/src/styles/variables.scss
